Move out test helper functions
